compile/scan: value, type IDs -> lower, upper IDs (#110)
Renaming value IDs to lower IDs and type IDs to upper IDs removes
the syntactic interpretation of the token (e.g., whether it is a
value or a type) from the lexical scanning. Additionally, it is less
confusing in those cases where, e.g., the upper IDs where also used
for things that where not types, like package names.