prototype: segregate function signatures (#148)
Function and FunctionType now use different signatures, namely
FunctionSignature and FunctionTypeSignature. The first one remains as
it was before, but FunctionTypeSignature no longer allows named
arguments (unlike Go). This avoids pushing a choice to the developer
that has no effect on the generated code (so it's mostly style) but
can have the effect of making types longer and potentially less
readable.

For interface methods and function declarations, we have chosen to
go with FunctionSignature and require named parameters, even if the
names aren't strictly necessary, for the sake of documentation and
readability.